@@ -1,7 +1,11 @@
-namespace Kritikos.Configuration.Persistence.Contracts.Behavioral
+namespace Kritikos.Configuration.Persistence.Contracts.Behavioral;
+
+/// <summary>
+/// Interface handling soft deletes to enable recovery of deleted entries.
+/// </summary>
+public interface ISoftDeletable
 {
-  public interface ISoftDeletable
-  {
-    public bool IsDeleted { get; set; }
-  }
+  public bool IsDeleted { get; set; }
+
+  public DateTime? DeletedAt { get; set; }
 }
