@@ -1,89 +1,96 @@
-namespace Kritikos.Configuration.Persistence.ConvertersTests {
-  using System.IO;
-  using System;
-  using Kritikos.Configuration.Persistence.Converters.IO;
-  using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
-  using Xunit;
+namespace Kritikos.Configuration.Persistence.ConvertersTests;
 
-  public class DirectoryInfoToStringTests {
-    private const string WindowsBase = @"C:\Windows\System32";
-    private const string WindowsRelative = @"drivers\etc";
+using System;
+using System.IO;
 
-    private const string WindowsPath = @"C:\Windows\System32\drivers\etc";
+using Kritikos.Configuration.Persistence.Converters.IO;
 
-    private const string LinuxBase = @"/srv/http";
-    private const string LinuxRelative = @"root";
-    private const string LinuxPath = @"/srv/http/root";
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
-    private static readonly ConverterMappingHints MappingHints = new (unicode: true);
+using Xunit;
 
-    [SkippableFact]
-    public void Relative_path_windows () {
-      Skip.If (true);
-      Skip.If (
-        Environment.OSVersion.Platform != PlatformID.Win32NT,
-        "Environment.OSVersion.Platform != PlatformID.Win32NT");
+public class DirectoryInfoToStringTests
+{
+  private const string WindowsBase = @"C:\Windows\System32";
+  private const string WindowsRelative = @"drivers\etc";
+  private const string WindowsPath = @"C:\Windows\System32\drivers\etc";
 
-      var converter =
-        new DirectoryInfoToStringConverter ('\\', new DirectoryInfo (WindowsBase), MappingHints);
+  private const string LinuxBase = "/srv/http";
+  private const string LinuxRelative = "root";
+  private const string LinuxPath = "/srv/http/root";
 
-      var file = converter.ConvertFromProvider (WindowsRelative) as DirectoryInfo;
-      var foo = converter.ConvertToProvider (file) as string;
+  private static readonly ConverterMappingHints MappingHints = new(unicode: true);
 
-      Assert.Equal (WindowsPath, file!.FullName);
-      Assert.Equal (WindowsRelative, foo);
-    }
+  [SkippableFact]
+  public void Relative_path_windows()
+  {
+    Skip.If(true);
+    Skip.If(
+      Environment.OSVersion.Platform != PlatformID.Win32NT,
+      "Environment.OSVersion.Platform != PlatformID.Win32NT");
 
-    [SkippableFact]
-    public void Absolute_path_windows () {
-      Skip.If (true);
-      Skip.If (
-        Environment.OSVersion.Platform != PlatformID.Win32NT,
-        "Environment.OSVersion.Platform != PlatformID.Win32NT");
+    var converter =
+      new DirectoryInfoToStringConverter('\\', new DirectoryInfo(WindowsBase), MappingHints);
 
-      var converter =
-        new DirectoryInfoToStringConverter ('\\', mappingHints : MappingHints);
+    var file = converter.ConvertFromProvider(WindowsRelative) as DirectoryInfo;
+    var foo = converter.ConvertToProvider(file) as string;
 
-      var file = converter.ConvertFromProvider (WindowsPath) as DirectoryInfo;
-      var foo = converter.ConvertToProvider (file) as string;
+    Assert.Equal(WindowsPath, file.FullName);
+    Assert.Equal(WindowsRelative, foo);
+  }
 
-      Assert.Equal (WindowsPath, file!.FullName);
-      Assert.Equal (WindowsRelative, foo);
-    }
+  [SkippableFact]
+  public void Absolute_path_windows()
+  {
+    Skip.If(true);
+    Skip.If(
+      Environment.OSVersion.Platform != PlatformID.Win32NT,
+      "Environment.OSVersion.Platform != PlatformID.Win32NT");
 
-    [SkippableFact]
-    public void Relative_path_linux () {
-      Skip.If (true);
-      Skip.If (
-        Environment.OSVersion.Platform == PlatformID.Win32NT,
-        "Environment.OSVersion.Platform == PlatformID.Win32NT");
+    var converter =
+      new DirectoryInfoToStringConverter('\\', mappingHints: MappingHints);
 
-      var converter =
-        new DirectoryInfoToStringConverter (
-          '/',
-          new DirectoryInfo (LinuxBase),
-          mappingHints : MappingHints);
+    var file = converter.ConvertFromProvider(WindowsPath) as DirectoryInfo;
+    var foo = converter.ConvertToProvider(file) as string;
 
-      var file = converter.ConvertFromProvider (LinuxRelative) as DirectoryInfo;
-      var foo = converter.ConvertToProvider (file) as string;
+    Assert.Equal(WindowsPath, file.FullName);
+    Assert.Equal(WindowsRelative, foo);
+  }
 
-      Assert.Equal (LinuxRelative, foo);
-    }
+  [SkippableFact]
+  public void Relative_path_linux()
+  {
+    Skip.If(true);
+    Skip.If(
+      Environment.OSVersion.Platform == PlatformID.Win32NT,
+      "Environment.OSVersion.Platform == PlatformID.Win32NT");
 
-    [SkippableFact]
-    public void Absolute_path_linux () {
-      Skip.If (true);
-      Skip.If (
-        Environment.OSVersion.Platform == PlatformID.Win32NT,
-        "Environment.OSVersion.Platform == PlatformID.Win32NT");
+    var converter =
+      new DirectoryInfoToStringConverter(
+        '/',
+        new DirectoryInfo(LinuxBase),
+        mappingHints: MappingHints);
 
-      var converter =
-        new DirectoryInfoToStringConverter ('/', mappingHints : MappingHints);
+    var file = converter.ConvertFromProvider(LinuxRelative) as DirectoryInfo;
+    var foo = converter.ConvertToProvider(file) as string;
 
-      var file = converter.ConvertFromProvider (LinuxPath) as DirectoryInfo;
-      var foo = converter.ConvertToProvider (file) as string;
+    Assert.Equal(LinuxRelative, foo);
+  }
 
-      Assert.Equal (LinuxPath, foo);
-    }
+  [SkippableFact]
+  public void Absolute_path_linux()
+  {
+    Skip.If(true);
+    Skip.If(
+      Environment.OSVersion.Platform == PlatformID.Win32NT,
+      "Environment.OSVersion.Platform == PlatformID.Win32NT");
+
+    var converter =
+      new DirectoryInfoToStringConverter('/', mappingHints: MappingHints);
+
+    var file = converter.ConvertFromProvider(LinuxPath) as DirectoryInfo;
+    var foo = converter.ConvertToProvider(file) as string;
+
+    Assert.Equal(LinuxPath, foo);
   }
 }
