@@ -1,89 +1,97 @@
-namespace Kritikos.Configuration.Persistence.ConvertersTests {
-  using System.IO;
-  using System;
-  using Kritikos.Configuration.Persistence.Converters.IO;
-  using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
-  using Xunit;
+namespace Kritikos.Configuration.Persistence.ConvertersTests;
 
-  public class FileInfoToStringTests {
-    private const string WindowsBase = @"C:\Windows\System32";
-    private const string WindowsRelative = @"drivers\etc\hosts";
+using System;
+using System.IO;
 
-    private const string WindowsPath = @"C:\Windows\System32\drivers\etc\hosts";
+using Kritikos.Configuration.Persistence.Converters.IO;
 
-    private const string LinuxBase = @"/srv/http";
-    private const string LinuxRelative = @"root/index.html";
-    private const string LinuxPath = @"/srv/http/root/index.html";
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
-    private static readonly ConverterMappingHints MappingHints = new (unicode: true);
+using Xunit;
 
-    [SkippableFact]
-    public void Relative_path_windows () {
-      Skip.If (true);
-      Skip.If (
-        Environment.OSVersion.Platform != PlatformID.Win32NT,
-        "Environment.OSVersion.Platform != PlatformID.Win32NT");
+public class FileInfoToStringTests
+{
+  private const string WindowsBase = @"C:\Windows\System32";
+  private const string WindowsRelative = @"drivers\etc\hosts";
 
-      var converter =
-        new FileInfoToStringConverter ('\\', new DirectoryInfo (WindowsBase), MappingHints);
+  private const string WindowsPath = @"C:\Windows\System32\drivers\etc\hosts";
 
-      var file = converter.ConvertFromProvider (WindowsRelative) as FileInfo;
-      var foo = converter.ConvertToProvider (file) as string;
+  private const string LinuxBase = "/srv/http";
+  private const string LinuxRelative = "root/index.html";
+  private const string LinuxPath = "/srv/http/root/index.html";
 
-      Assert.Equal (WindowsPath, file!.FullName);
-      Assert.Equal (WindowsRelative, foo);
-    }
+  private static readonly ConverterMappingHints MappingHints = new(unicode: true);
 
-    [SkippableFact]
-    public void Absolute_path_windows () {
-      Skip.If (true);
-      Skip.If (
-        Environment.OSVersion.Platform != PlatformID.Win32NT,
-        "Environment.OSVersion.Platform != PlatformID.Win32NT");
+  [SkippableFact]
+  public void Relative_path_windows()
+  {
+    Skip.If(true);
+    Skip.If(
+      Environment.OSVersion.Platform != PlatformID.Win32NT,
+      "Environment.OSVersion.Platform != PlatformID.Win32NT");
 
-      var converter =
-        new FileInfoToStringConverter ('\\', mappingHints : MappingHints);
+    var converter =
+      new FileInfoToStringConverter('\\', new DirectoryInfo(WindowsBase), MappingHints);
 
-      var file = converter.ConvertFromProvider (WindowsPath) as FileInfo;
-      var foo = converter.ConvertToProvider (file) as string;
+    var file = converter.ConvertFromProvider(WindowsRelative) as FileInfo;
+    var foo = converter.ConvertToProvider(file) as string;
 
-      Assert.Equal (WindowsPath, file!.FullName);
-      Assert.Equal (WindowsRelative, foo);
-    }
+    Assert.Equal(WindowsPath, file.FullName);
+    Assert.Equal(WindowsRelative, foo);
+  }
 
-    [SkippableFact]
-    public void Relative_path_nix () {
-      Skip.If (true);
-      Skip.If (
-        Environment.OSVersion.Platform == PlatformID.Win32NT,
-        "Environment.OSVersion.Platform == PlatformID.Win32NT");
+  [SkippableFact]
+  public void Absolute_path_windows()
+  {
+    Skip.If(true);
+    Skip.If(
+      Environment.OSVersion.Platform != PlatformID.Win32NT,
+      "Environment.OSVersion.Platform != PlatformID.Win32NT");
 
-      var converter =
-        new FileInfoToStringConverter (
-          '/',
-          new DirectoryInfo (LinuxBase),
-          mappingHints : MappingHints);
+    var converter =
+      new FileInfoToStringConverter('\\', mappingHints: MappingHints);
 
-      var file = converter.ConvertFromProvider (LinuxRelative) as FileInfo;
-      var foo = converter.ConvertToProvider (file) as string;
+    var file = converter.ConvertFromProvider(WindowsPath) as FileInfo;
+    var foo = converter.ConvertToProvider(file) as string;
 
-      Assert.Equal (LinuxRelative, foo);
-    }
+    Assert.Equal(WindowsPath, file.FullName);
+    Assert.Equal(WindowsRelative, foo);
+  }
 
-    [SkippableFact]
-    public void Absolute_path_nix () {
-      Skip.If (true);
-      Skip.If (
-        Environment.OSVersion.Platform == PlatformID.Win32NT,
-        "Environment.OSVersion.Platform == PlatformID.Win32NT");
+  [SkippableFact]
+  public void Relative_path_nix()
+  {
+    Skip.If(true);
+    Skip.If(
+      Environment.OSVersion.Platform == PlatformID.Win32NT,
+      "Environment.OSVersion.Platform == PlatformID.Win32NT");
 
-      var converter =
-        new FileInfoToStringConverter ('/', mappingHints : MappingHints);
+    var converter =
+      new FileInfoToStringConverter(
+        '/',
+        new DirectoryInfo(LinuxBase),
+        mappingHints: MappingHints);
 
-      var file = converter.ConvertFromProvider (LinuxPath) as FileInfo;
-      var foo = converter.ConvertToProvider (file) as string;
+    var file = converter.ConvertFromProvider(LinuxRelative) as FileInfo;
+    var foo = converter.ConvertToProvider(file) as string;
 
-      Assert.Equal (LinuxPath, foo);
-    }
+    Assert.Equal(LinuxRelative, foo);
+  }
+
+  [SkippableFact]
+  public void Absolute_path_nix()
+  {
+    Skip.If(true);
+    Skip.If(
+      Environment.OSVersion.Platform == PlatformID.Win32NT,
+      "Environment.OSVersion.Platform == PlatformID.Win32NT");
+
+    var converter =
+      new FileInfoToStringConverter('/', mappingHints: MappingHints);
+
+    var file = converter.ConvertFromProvider(LinuxPath) as FileInfo;
+    var foo = converter.ConvertToProvider(file) as string;
+
+    Assert.Equal(LinuxPath, foo);
   }
 }
